Moved the dob to the licence summary is preparation for contact summary rework
