+++ packages/gafl-webapp-service/src/pages/summary/contact-summary/contact-summary.njk
@@ -33,24 +33,6 @@
     },
     {
       key: {
-        text: "Date of birth"
-      },
-      value: {
-        text: data.birthDateStr
-      },
-      actions: {
-        items: [
-          {
-            href: data.uri.dateOfBirth,
-            text: "Change",
-            visuallyHiddenText: "date of birth",
-            attributes: { id: 'change-birth-date' }
-          }
-        ]
-      }
-    },
-    {
-      key: {
         text: "Address"
       },
       value: {
