+++ packages/gafl-webapp-service/src/pages/summary/licence-summary/licence-summary.njk
@@ -74,23 +74,43 @@
 {% endif %}
 
 {{ licenseDetailsSummaryRows.push({
-  key: {
-    text: "Valid from"
-  },
-  value: {
-    text: '30 minutes after payment' if data.permission.licenceToStart === 'after-payment' else data.startTimeString
-  },
-  actions: {
+      key: {
+        text: "Valid from"
+      },
+      value: {
+        text: '30 minutes after payment' if data.permission.licenceToStart === 'after-payment' else data.startTimeString
+      },
+      actions: {
+        items: [
+          {
+            href: data.uri.licenceToStart,
+            text: "Change",
+            visuallyHiddenText: "valid from",
+            attributes: { id: 'change-licence-to-start' }
+          }
+        ]
+      }
+}) }}
+
+{{ licenseDetailsSummaryRows.push({
+    key: {
+    text: "Date of birth"
+    },
+    value: {
+    text: data.birthDateStr
+    },
+    actions: {
     items: [
       {
-        href: data.uri.licenceToStart,
+        href: data.uri.dateOfBirth,
         text: "Change",
-        visuallyHiddenText: "valid from",
-        attributes: { id: 'change-licence-to-start' }
+        visuallyHiddenText: "date of birth",
+        attributes: { id: 'change-birth-date' }
       }
     ]
-  }
+    }
 }) }}
+
 
 {% if data.hasJunior %}
     {{ licenseDetailsSummaryRows.push({
